Move Journal related classes to its own package
