+++ src/main/scala/akka/persistence/jdbc/dao/bytea/journal/JournalQueries.scala
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package akka.persistence.jdbc.dao.bytea
+package akka.persistence.jdbc.dao.bytea.journal
 
-import akka.persistence.jdbc.dao.bytea.JournalTables.{ JournalDeletedToRow, JournalRow }
-import akka.persistence.jdbc.config.{ DeletedToTableConfiguration, JournalTableConfiguration }
-import akka.persistence.jdbc.serialization.{ SerializationResult, Serialized }
+import akka.persistence.jdbc.config.{DeletedToTableConfiguration, JournalTableConfiguration}
+import akka.persistence.jdbc.dao.bytea.journal.JournalTables.{JournalDeletedToRow, JournalRow}
 import slick.driver.JdbcProfile
 
 class JournalQueries(val profile: JdbcProfile, override val journalTableCfg: JournalTableConfiguration, override val deletedToTableCfg: DeletedToTableConfiguration) extends JournalTables {
